Back to collision!
Everything works and collision has started being worked on. Still
doesn't fall after player goes off the blocks but quinn may have come up
with a solution. Right now the player stops listening to gravity as soon
as collision returns as true. He suggested making gravity a constant
even if collision is true and then just make the player move out of the
box when he collides. Need to make that change and see what happens.
After that move on to making the window move with the player. Still
going to have to think on how i'm going to do that one though...